Add bowl-slosh test
