+++ examples/tsunami/bowl-slosh/regression_data/claw_git_status.txt
