+++ examples/tsunami/bowl-slosh/regression_data/gauge00001.txt
