+++ examples/tsunami/bowl-slosh/regression_data/regression_data_fgmax.txt
